Cleaned up code according to compiler warnings

+++ src/packet/Packet.cpp
@@ -20,12 +20,12 @@ void Packet::writeHeader(MemoryStream& ms, uint8_t type)
   if (!m_started) {
     throw std::logic_error("Packet must be started.");
   }
-  size_t pos = ms.seekP(m_pos, MemoryStream::Direction::Begin);
+  size_t pos = ms.seekP(static_cast<ssize_t>(m_pos), MemoryStream::Direction::Begin);
   m_pos += HeaderSize;
   if (pos < m_pos) {
     throw std::logic_error("Violation of writing order in the stream");
   }
   ms.writeUInt8(type);
-  ms.seekP(pos, MemoryStream::Direction::Begin);
+  ms.seekP(static_cast<ssize_t>(pos), MemoryStream::Direction::Begin);
   m_started = false;
 }
